added square root of 2 and fibonacci. seems to work so far.
does not work with leibniz formula for PI (why?) it converges slowly anyway
but the even and odd convergents do not bracket the true value ... which is disturbing.

needs more testing. like a real, legit test suite.

instead of having ad-hoc methods, PseudoReal and generalizedCF
should actually implement existing typeclasses.

generalizedCFs are not memoized, so it takes exponential time
to get a particular convergent (I think). scary.

add a sum and product function that split up the "allowed error" evenly
between the items. Actually product might be tricky because we have to do
iterative approximation.